requirements.txt
scipy
nltk>=3.0.5,<3.5  # TODO: change when fixed: https://bitbucket.org/mrabarnett/mrab-regex/issues/369/please-provide-macos-wheels-on-pypi
scikit-learn
numpy
python-dateutil<3.0.0  # denpendency for botocore
gensim>=0.12.3  # LDA's show topics unified in 0.12.3
setuptools-git
Orange3 >=3.25.0
tweepy
beautifulsoup4
simhash >=1.11
wikipedia
pdfminer3k>=1.3.1
odfpy>=1.3.5
docx2txt>=0.6
lxml
biopython # Enables Pubmed widget.
ufal.udpipe >=1.2.0.3
